refactor(bytes-util): modules and coverage
