+++ crates/bytes-util/src/cow/string/serde.rs
@@ -1,74 +1,4 @@
-use super::{BytesCow, StringCow};
-
-impl serde::Serialize for BytesCow<'_> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_bytes(self.as_bytes())
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for BytesCow<'de> {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        struct BytesCowVisitor;
-
-        impl<'de> serde::de::Visitor<'de> for BytesCowVisitor {
-            type Value = BytesCow<'de>;
-
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("a byte slice")
-            }
-
-            fn visit_byte_buf<E>(self, v: Vec<u8>) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(BytesCow::from_vec(v))
-            }
-
-            fn visit_bytes<E>(self, v: &[u8]) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(BytesCow::from_vec(v.to_vec()))
-            }
-
-            fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(BytesCow::from_vec(v.as_bytes().to_vec()))
-            }
-
-            fn visit_string<E>(self, v: String) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(BytesCow::from_vec(v.into_bytes()))
-            }
-
-            fn visit_borrowed_bytes<E>(self, v: &'de [u8]) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(BytesCow::from_slice(v))
-            }
-
-            fn visit_borrowed_str<E>(self, v: &'de str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(BytesCow::from_slice(v.as_bytes()))
-            }
-        }
-
-        deserializer.deserialize_bytes(BytesCowVisitor)
-    }
-}
+use super::StringCow;
 
 impl serde::Serialize for StringCow<'_> {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
@@ -146,7 +76,6 @@ impl<'a, E> StringCowDeserializer<'a, E> {
     }
 }
 
-#[cfg(feature = "serde")]
 impl<'de, E> serde::de::Deserializer<'de> for StringCowDeserializer<'de, E>
 where
     E: serde::de::Error,
